Move cliutils outside of cli submodule to avoid initializing cli
