+++ newrelic_lambda_cli/cli/layers.py
@@ -4,7 +4,7 @@ import boto3
 import click
 
 from .. import layers, permissions
-from .cliutils import done
+from ..cliutils import done
 from .decorators import add_options, AWS_OPTIONS
