+++ newrelic_lambda_cli/cli/subscriptions.py
@@ -2,7 +2,7 @@ import boto3
 import click
 
 from .. import permissions, subscriptions
-from .cliutils import done
+from ..cliutils import done
 from .decorators import add_options, AWS_OPTIONS
